[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17700 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,8 +5,8 @@ public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		
-		int A = sc.nextInt();
-		int B = sc.nextInt();
+		long A = sc.nextInt();
+		long B = sc.nextInt();
 		
 		System.out.println((A + B) * (A - B));
 	}
